Reworked 'get_birthdays' function: now it has O(n) complexity while iterating through records. Relocated 'DictSortable(UserDict)' class to AddressBook.py and deleted get_birthdays_project.py
